+++ packages/components/src/components/atoms/form/DatePickerWithoutError.stories.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react'
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 
-import DatePickerComponent from './DatePicker'
-import DatePickerError from './DatePickerError'
+import DatePickerWithoutErrorComponent from './DatePickerWithoutError'
+import DatePicker from './DatePicker'
 
 export default {
   title: 'Atoms/form/DatePicker',
-  component: DatePickerComponent,
+  component: DatePickerWithoutErrorComponent,
   argTypes: {
     color: {
       options: ['none', 'success', 'error'],
@@ -31,16 +31,24 @@ export default {
       control: 'hidden',
     },
   },
-} as ComponentMeta<typeof DatePickerComponent>
+} as ComponentMeta<typeof DatePickerWithoutErrorComponent>
 
-const Template: ComponentStory<typeof DatePickerComponent> = (args) => {
+const Template: ComponentStory<typeof DatePickerWithoutErrorComponent> = (
+  args
+) => {
   const [value, setValue] = useState<Date | null>(null)
 
   function onChange(value: Date | null): void {
     setValue(value)
   }
 
-  return <DatePickerComponent {...args} value={value} onChange={onChange} />
+  return (
+    <DatePickerWithoutErrorComponent
+      {...args}
+      value={value}
+      onChange={onChange}
+    />
+  )
 }
 
 export const Default = Template.bind({})
@@ -59,14 +67,14 @@ Default.args = {
   transparent: false,
 }
 
-const FormErrorTemplate: ComponentStory<typeof DatePickerError> = (args) => {
+const FormErrorTemplate: ComponentStory<typeof DatePicker> = (args) => {
   const [value, setValue] = useState<Date | null>(null)
 
   function onChange(value: Date | null): void {
     setValue(value)
   }
 
-  return <DatePickerError {...args} value={value} onChange={onChange} />
+  return <DatePicker {...args} value={value} onChange={onChange} />
 }
 
 export const WithError = FormErrorTemplate.bind({})
