+++ packages/components/src/components/index.ts
@@ -5,8 +5,8 @@ export { default as Breadcrumbs } from './atoms/breadcrumb/Breadcrumbs'
 export { default as Chip } from './atoms/Chip/Chip'
 export { default as CloseComponent } from './atoms/closeComponent/CloseComponent'
 export { default as Checkbox } from './atoms/form/CheckboxWithoutError'
-export { default as DatePicker } from './atoms/form/DatePicker'
-export { default as DatePickerError } from './atoms/form/DatePickerError'
+export { default as DatePicker } from './atoms/form/DatePickerWithoutError'
+export { default as DatePickerError } from './atoms/form/DatePicker'
 export { default as DoubleDatePicker } from './atoms/form/DoubleDatePicker'
 export { default as DoubleDatePickerError } from './atoms/form/DoubleDatePickerError'
 export { default as DropDown } from './atoms/form/DropDown'
